PanelEditorNext: Fixes issue with tableView and query updates
