+++ packages/grafana-data/src/types/options.ts
@@ -49,6 +49,10 @@ export interface OptionEditorConfig<TOptions, TSettings = any, TValue = any> {
   defaultValue?: TValue;
 
   /**
+   *
+   */
+
+  /**
    * Function that enables configuration of when option editor should be shown based on current panel option properties.
    */
   showIf?: (currentOptions: TOptions, data?: DataFrame[], annotations?: DataFrame[]) => boolean | undefined;
