+++ public/app/features/dashboard-scene/panel-edit/PanelEditNext/VizAndDataPaneNext.test.tsx
@@ -28,6 +28,10 @@ jest.mock('./QueryEditor/Sidebar/Sidebar', () => ({
   Sidebar: () => <div data-testid="query-editor-sidebar" />,
 }));
 
+jest.mock('../PanelEditPanelWrapper', () => ({
+  PanelEditPanelWrapper: () => <div data-testid="panel-viz" />,
+}));
+
 // Minimal mock so instanceof checks in VizAndDataPaneNext work without scene setup
 jest.mock('./PanelDataPaneNext', () => ({
   PanelDataPaneNext: class {
@@ -40,9 +44,11 @@ const MockPanelComponent = () => <div data-testid="panel-viz" />;
 function buildMockLayout(dataPane?: PanelDataPane | PanelDataPaneNext) {
   return {
     scene: {
-      panelToShow: { Component: MockPanelComponent },
+      panel: { Component: MockPanelComponent },
+      tableView: { Component: MockPanelComponent },
       controls: null,
       dataPane,
+      dashboard: {},
     },
     layout: {
       sidebarSize: SidebarSize.Mini,
